Added check_timestamps property to videoaggregator
Added a property to videoagrregator to control if the timestamps of all
sink pads should match in order for a frame to be produced (will ignore
		the fps and won't duplicate any buffers).
This was needed for the metadatamuxer.

Improvements when running with check_timestamps=true (QOS problems)

If we are producing frames based only on the matching timestamps
(and therefore dont care about the fps) we need to keep the nframes=0
and adjust the segment based on the produced frames instead.

Otherwise the QOS management system starts complaining and dropping
frames (if frames are not produced regularly, fx. a muxer after a
stitcher)

[Merged with Defined-GST_FLOW_VIDEO_AGGREGATOR_DROP_FRAME-GstFlow and
Added-function-to-reset-segment-position-in-videoagg added fix from msb]

Credit-to:  Marianna Smidth Buschle <msb@qtec.com>